// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/verificationDatetimeSectionResponseBean.ts
import {VerificationDatetimeItemBean} from "./verificationDatetimeItemBean";
import {VerificationDatetimeTotalBean} from "./verificationDatetimeTotalBean";

export class VerificationDatetimeSectionResponseBean{
  token?: string;
  start?: VerificationDatetimeItemBean;
  end?: VerificationDatetimeItemBean;
  total?: VerificationDatetimeTotalBean;

}